hoist action to change qs to component
