+++ web/src/components/CountryDistribution/useCountryQuery.tsx
@@ -1,10 +1,9 @@
 import React, { useState, useEffect, useMemo } from 'react'
 import { useRouter } from 'next/router'
-import { stringify } from 'querystring'
 import { Region, ClusterState, getPerCountryData, CountryDistribution } from 'src/io/getPerCountryData'
 import { Places } from 'src/io/getPlaces'
 
-import { getRegionBySelectedCountries, getCurriedClustersBySelectedClusters, getCurrentQs } from './utils'
+import { getRegionBySelectedCountries, getCurriedClustersBySelectedClusters } from './utils'
 
 export type ParsedUrlQuery = string | string[] | undefined
 
@@ -28,7 +27,6 @@ export const useCountryAndClusterQuery = (): {
   }
   state: {
     region: Region
-    setRegion: (region: Region) => void
     places: Places
     setPlaces: React.Dispatch<React.SetStateAction<Places>>
     countryDistributions: CountryDistribution[]
@@ -61,23 +59,6 @@ export const useCountryAndClusterQuery = (): {
   const [places, setPlaces] = useState<Places>(initialPlaces)
   const [currentClusters, setClusters] = useState<ClusterState>(getClustersBySelectedClusters(selectedClusters))
 
-  const setCurrentRegionAsQs = (nextRegion: Region): void => {
-    if (currentRegion === nextRegion) {
-      return
-    }
-    const fullPath = `${router.basePath}${router.pathname}`
-    const nextRegionQs = { ...getCurrentQs(router) }
-
-    if (nextRegion === Region.WORLD) {
-      delete nextRegionQs.countries
-    } else if (nextRegion === Region.UNITED_STATES) {
-      nextRegionQs.countries = 'usa'
-    } else if (nextRegion === Region.SWITZERLAND) {
-      nextRegionQs.countries = 'switzerland'
-    }
-    return void router.replace(`${fullPath}?${stringify(nextRegionQs)}`)
-  }
-
   useEffect(() => {
     setPlaces(initialPlaces)
   }, [initialPlaces, setPlaces])
@@ -101,7 +82,6 @@ export const useCountryAndClusterQuery = (): {
     },
     state: {
       region: currentRegion,
-      setRegion: setCurrentRegionAsQs,
       places,
       setPlaces,
       countryDistributions,
